requirements.txt
Flask==1.0.2
pyjwt
cryptography
requests
google-cloud-firestore